+++ backend/src/domain/album/dto/albumInfo.ts
