+++ backend/src/domain/album/swagger/createAlbum.swagger.ts
@@ -1,6 +1,6 @@
 import { applyDecorators } from "@nestjs/common";
 import { ApiResponse } from "@nestjs/swagger";
-import { CreateAlbumResponseDto } from "src/dto/album/createAlbumResponse.dto";
+import { CreateAlbumResponseDto } from "src/domain/album/dto/createAlbumResponse.dto";
 
 export function SwaggerCreateAlbum() {
   return applyDecorators(ApiResponse({ type: CreateAlbumResponseDto, status: 200 }));
