+++ backend/src/domain/group/dto/attendGroupRequest.dto.ts
