+++ backend/src/domain/group/dto/createGroupResponse.dto.ts
