+++ backend/src/domain/group/dto/getAlbumsResponse.dto.ts
