+++ backend/src/domain/group/dto/getHashTagsResponse.dto.ts
