+++ backend/src/domain/group/dto/groupInfo.ts
