+++ backend/src/domain/group/dto/updateGroupInfoRequest.dto.ts
