+++ backend/src/domain/group/dto/updateGroupInfoResponse.dto.ts
