+++ backend/src/domain/image/dto/imageInfo.ts
