+++ backend/src/domain/post/post.service.ts
@@ -5,11 +5,11 @@ import { InjectRepository } from "@nestjs/typeorm";
 import { PostRepository } from "./post.repository";
 import { UserRepository } from "src/domain/user/user.repository";
 import { AlbumRepository } from "src/domain/album/album.repository";
-import { CreatePostRequestDto } from "src/dto/post/createPostRequest.dto";
-import { GetPostInfoResponseDto } from "src/dto/post/getPostInfoResponse.dto";
-import { UpdatePostInfoRequestDto } from "src/dto/post/updatePostInfoRequest.dto";
-import { ShiftPostRequestDto } from "src/dto/post/shiftPostRequest.dto";
-import { GetSearchPostResponse } from "src/dto/post/getSearchPostResponse.dto";
+import { CreatePostRequestDto } from "src/domain/post/dto/createPostRequest.dto";
+import { GetPostInfoResponseDto } from "src/domain/post/dto/getPostInfoResponse.dto";
+import { UpdatePostInfoRequestDto } from "src/domain/post/dto/updatePostInfoRequest.dto";
+import { ShiftPostRequestDto } from "src/domain/post/dto/shiftPostRequest.dto";
+import { GetSearchPostResponse } from "src/domain/post/dto/getSearchPostResponse.dto";
 import { AlbumService } from "src/domain/album/album.service";
 import { HashTagService } from "src/domain/hashtag/hashtag.service";
 import { HashTag } from "src/domain/hashtag/hashtag.entity";
