+++ backend/src/domain/user/dto/getGroupsResponse.dto.ts
