+++ backend/src/domain/user/spec/user.service.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from "@nestjs/testing";
 import { getRepositoryToken } from "@nestjs/typeorm";
-import { UserInfoResponseDto } from "src/dto/user/userInfoResponse.dto";
+import { UserInfoResponseDto } from "src/domain/user/dto/userInfoResponse.dto";
 import { Group } from "src/domain/group/group.entity";
 import { User } from "../user.entity";
 import { UserRepository } from "../user.repository";
