+++ backend/src/domain/user/user.service.ts
@@ -1,14 +1,14 @@
 import { Injectable, NotFoundException } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
-import { UserInfo } from "src/dto/user/userInfo.dto";
+import { UserInfo } from "src/domain/user/dto/userInfo.dto";
 import { User } from "./user.entity";
 import { UserRepository } from "./user.repository";
 import { CustomFile } from "src/custom/myFile/customFile";
-import { UserInfoResponseDto } from "src/dto/user/userInfoResponse.dto";
-import { UpdateUserInfoRequestDto } from "src/dto/user/updateUserInfoRequest.dto";
-import { UpdateGroupOrderRequestDto } from "src/dto/user/updateGroupOrderRequest.dto";
-import { GetGroupsResponseDto } from "src/dto/user/getGroupsResponse.dto";
-import { UpdateUserInfoResponseDto } from "src/dto/user/updateUserInfoResponse.dto";
+import { UserInfoResponseDto } from "src/domain/user/dto/userInfoResponse.dto";
+import { UpdateUserInfoRequestDto } from "src/domain/user/dto/updateUserInfoRequest.dto";
+import { UpdateGroupOrderRequestDto } from "src/domain/user/dto/updateGroupOrderRequest.dto";
+import { GetGroupsResponseDto } from "src/domain/user/dto/getGroupsResponse.dto";
+import { UpdateUserInfoResponseDto } from "src/domain/user/dto/updateUserInfoResponse.dto";
 import { UpdateResult } from "typeorm";
 import { getImageUrl } from "src/common/imageUrl";
 import { ArrayToObject, reArrange } from "src/common/changeObject";
